Upgraded PHPUnit to version 3.3.9

+++ lib/PHPUnit/Extensions/PhptTestCase.php
@@ -39,16 +39,25 @@
  * @author     Sebastian Bergmann <sb@sebastian-bergmann.de>
  * @copyright  2002-2008 Sebastian Bergmann <sb@sebastian-bergmann.de>
  * @license    http://www.opensource.org/licenses/bsd-license.php  BSD License
- * @version    SVN: $Id: PhptTestCase.php 1985 2007-12-26 18:11:55Z sb $
+ * @version    SVN: $Id: PhptTestCase.php 4219 2008-12-10 08:56:57Z sb $
  * @link       http://www.phpunit.de/
  * @since      File available since Release 3.1.4
  */
 
+if (PHPUnit_Util_Filesystem::fileExistsInIncludePath('PEAR/RunTest.php')) {
+    $currentErrorReporting = error_reporting(E_ERROR | E_WARNING | E_PARSE);
+    PHPUnit_Util_Filesystem::collectStart();
+    require_once 'PEAR/RunTest.php';
+    error_reporting($currentErrorReporting);
+
+    foreach (PHPUnit_Util_Filesystem::collectEnd() as $blacklistedFile) {
+        PHPUnit_Util_Filter::addFileToFilter($blacklistedFile, 'PHPUNIT');
+    }
+}
+
 require_once 'PHPUnit/Framework.php';
 require_once 'PHPUnit/Extensions/PhptTestCase/Logger.php';
 require_once 'PHPUnit/Util/Filter.php';
-
-@include_once 'PEAR/RunTest.php';
 
 PHPUnit_Util_Filter::addFileToFilter(__FILE__, 'PHPUNIT');
 
@@ -60,20 +69,24 @@ PHPUnit_Util_Filter::addFileToFilter(__FILE__, 'PHPUNIT');
  * @author     Sebastian Bergmann <sb@sebastian-bergmann.de>
  * @copyright  2002-2008 Sebastian Bergmann <sb@sebastian-bergmann.de>
  * @license    http://www.opensource.org/licenses/bsd-license.php  BSD License
- * @version    Release: 3.2.9
+ * @version    Release: 3.3.9
  * @link       http://www.phpunit.de/
  * @since      Class available since Release 3.1.4
  */
-class PHPUnit_Extensions_PhptTestCase implements PHPUnit_Framework_Test
+class PHPUnit_Extensions_PhptTestCase implements PHPUnit_Framework_Test, PHPUnit_Framework_SelfDescribing
 {
     /**
      * The filename of the .phpt file.
      *
      * @var    string
-     * @access protected
      */
     protected $filename;
 
+    /**
+     * Options for PEAR_RunTest.
+     *
+     * @var    array
+     */
     protected $options = array();
 
     /**
@@ -82,7 +95,6 @@ class PHPUnit_Extensions_PhptTestCase implements PHPUnit_Framework_Test
      * @param  string $filename
      * @param  array  $options Array with ini settings for the php instance run,
      *                         key being the name if the setting, value the ini value.
-     * @access public
      */
     public function __construct($filename, $options = array())
     {
@@ -111,7 +123,6 @@ class PHPUnit_Extensions_PhptTestCase implements PHPUnit_Framework_Test
      * Counts the number of test cases executed by run(TestResult result).
      *
      * @return integer
-     * @access public
      */
     public function count()
     {
@@ -125,12 +136,19 @@ class PHPUnit_Extensions_PhptTestCase implements PHPUnit_Framework_Test
      * @param  array $options Array with ini settings for the php instance run,
      *                        key being the name if the setting, value the ini value.
      * @return PHPUnit_Framework_TestResult
-     * @access public
      */
     public function run(PHPUnit_Framework_TestResult $result = NULL, $options = array())
     {
         if (!class_exists('PEAR_RunTest', FALSE)) {
             throw new RuntimeException('Class PEAR_RunTest not found.');
+        }
+
+        if (isset($GLOBALS['_PEAR_destructor_object_list']) &&
+            is_array($GLOBALS['_PEAR_destructor_object_list']) &&
+            !empty($GLOBALS['_PEAR_destructor_object_list'])) {
+            $pearDestructorObjectListCount = count($GLOBALS['_PEAR_destructor_object_list']);
+        } else {
+            $pearDestructorObjectListCount = 0;
         }
 
         if ($result === NULL) {
@@ -141,19 +159,19 @@ class PHPUnit_Extensions_PhptTestCase implements PHPUnit_Framework_Test
             throw new InvalidArgumentException;
         }
 
-        $options = array_merge($options, $this->options);
-
         $coverage = $result->getCollectCodeCoverageInformation();
+        $options  = array_merge($options, $this->options);
 
         if ($coverage) {
-            $options = array('coverage' => TRUE);
+            $options['coverage'] = TRUE;
         } else {
-            $options = array();
+            $options['coverage'] = FALSE;
         }
 
-        $runner = new PEAR_RunTest(new PHPUnit_Extensions_PhptTestCase_Logger, $options);
+        $currentErrorReporting = error_reporting(E_ERROR | E_WARNING | E_PARSE);
+        $runner                = new PEAR_RunTest(new PHPUnit_Extensions_PhptTestCase_Logger, $options);
 
-        if ($coverage){
+        if ($coverage) {
             $runner->xdebug_loaded = TRUE;
         } else {
             $runner->xdebug_loaded = FALSE;
@@ -162,9 +180,9 @@ class PHPUnit_Extensions_PhptTestCase implements PHPUnit_Framework_Test
         $result->startTest($this);
 
         PHPUnit_Util_Timer::start();
-        $buffer = $runner->run($this->filename, $options);
-        $time = PHPUnit_Util_Timer::stop();
-
+        $buffer       = $runner->run($this->filename, $options);
+        $time         = PHPUnit_Util_Timer::stop();
+        error_reporting($currentErrorReporting);
         $base         = basename($this->filename);
         $path         = dirname($this->filename);
         $coverageFile = $path . DIRECTORY_SEPARATOR . str_replace('.phpt', '.xdebug', $base);
@@ -175,11 +193,11 @@ class PHPUnit_Extensions_PhptTestCase implements PHPUnit_Framework_Test
         $phpFile      = $path . DIRECTORY_SEPARATOR . str_replace('.phpt', '.php', $base);
 
         if (is_object($buffer) && $buffer instanceof PEAR_Error) {
-            $result->addError( 
-              $this, 
+            $result->addError(
+              $this,
               new RuntimeException($buffer->getMessage()),
-              $time 
-            ); 
+              $time
+            );
         }
 
         else if ($buffer == 'SKIPPED') {
@@ -191,9 +209,7 @@ class PHPUnit_Extensions_PhptTestCase implements PHPUnit_Framework_Test
               $this,
               PHPUnit_Framework_ComparisonFailure::diffEqual(
                 file_get_contents($expFile),
-                file_get_contents($outFile),
-                FALSE,
-                $this->getName()
+                file_get_contents($outFile)
               ),
               $time
             );
@@ -205,20 +221,27 @@ class PHPUnit_Extensions_PhptTestCase implements PHPUnit_Framework_Test
             }
         }
 
-        if ($coverage) {
+        if ($coverage && file_exists($coverageFile)) {
             eval('$coverageData = ' . file_get_contents($coverageFile) . ';');
             unset($coverageData[$phpFile]);
 
-            $codeCoverageInformation = array(
-              'test'  => $this,
-              'files' => $coverageData
-            );
-
-            $result->appendCodeCoverageInformation($this, $codeCoverageInformation);
+            $result->appendCodeCoverageInformation($this, $coverageData);
             unlink($coverageFile);
         }
 
         $result->endTest($this, $time);
+
+        // Do not invoke PEAR's destructor mechanism for PHP 4
+        // as it raises an E_STRICT.
+        if ($pearDestructorObjectListCount == 0) {
+            unset($GLOBALS['_PEAR_destructor_object_list']);
+        } else {
+            $count = count($GLOBALS['_PEAR_destructor_object_list']) - $pearDestructorObjectListCount;
+
+            for ($i = 0; $i < $count; $i++) {
+                array_pop($GLOBALS['_PEAR_destructor_object_list']);
+            }
+        }
 
         return $result;
     }
@@ -227,7 +250,6 @@ class PHPUnit_Extensions_PhptTestCase implements PHPUnit_Framework_Test
      * Returns the name of the test case.
      *
      * @return string
-     * @access public
      */
     public function getName()
     {
@@ -238,7 +260,6 @@ class PHPUnit_Extensions_PhptTestCase implements PHPUnit_Framework_Test
      * Returns a string representation of the test case.
      *
      * @return string
-     * @access public
      */
     public function toString()
     {
